requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy
pyyaml
scikit-image
xlrd
wandb